refactor:Vice Dean's assignment#2.1
