+++ Hansang/Assignment/src/main/java/com/nhnacademy/Assignment/SearchEngineShell.java
@@ -2,22 +2,15 @@ package com.nhnacademy.Assignment;
 
 import static java.lang.CharSequence.compare;
 
-import com.nhnacademy.Assignment.resource.Inputshell;
-import java.util.ArrayList;
 import java.util.List;
 
 
 public class SearchEngineShell {
-
-    List<Movie> searchedData = new ArrayList<>();
-
     public void loadEngine() {
-        String path = "./src/main/java/com/nhnacademy/Assignment/resource/Movie.csv";
-
         List<Movie> movieList;
 
         try {
-            movieList = scanFile.listingMovies(path);
+            movieList = scanFile.listingMovies("Movie.csv");
             movieList.sort((title1, title2) -> compare(title1.title(), title2.title()));
             searchName(movieList);
         } catch (Exception e) {
@@ -30,7 +23,7 @@ public class SearchEngineShell {
         while (true) {
             System.out.println("영화 검색 엔진입니다. 원하시는 영화를 검색해주세요..(종료하시려면 \"stop\"를 입력해주세요)");
             String movieName = Inputshell.inputFromShell(">> ");
-            List<Movie> searchResult;
+            int searchResult;
             if (movieName.toUpperCase().trim().equals("STOP")) {
                 Inputshell.scannerClose();
                 return;
@@ -38,20 +31,21 @@ public class SearchEngineShell {
             if (movieName.isEmpty()) {
                 continue;
             }
-            searchedData.clear();
             searchResult = binSearch(movieName, movieList);
 
-            showResult(searchResult);
+            showResult(searchResult, movieList);
+        }
+    }
+    
+    private void showResult(int searchResult, List<Movie> movieList) {
+        if (searchResult == -1) {
+            System.out.println("no data");
+        } else {
+            System.out.println(movieList.get(searchResult).toString());
         }
     }
 
-    private void showResult(List<Movie> searchResult) {
-        for (int i = 0; i < searchResult.size(); i++) {
-            System.out.println(searchResult.get(i).toString());
-        }
-    }
-
-    private List<Movie> binSearch(String movieName, List<Movie> movieList) {
+    private int binSearch(String movieName, List<Movie> movieList) {
         int start = 0;
         int end = movieList.size() - 1;
         int node;
@@ -65,21 +59,14 @@ public class SearchEngineShell {
             compareNum = movieName.compareTo(movieList.get(node).title());
 
             if (compareNum == 0) {
-                searchedData.add(movieList.get(node));
-                if (start == end) {
-                    break;
-                }
+                return node;
             } else if (compareNum < 0) {
                 end = node - 1;
             } else {
                 start = node + 1;
             }
         }
-        if (searchedData.isEmpty()) {
-            System.out.println("no data");
-        }
-
-        return searchedData;
+        return -1;
     }
 
 }
